Split up Query test
