+++ test/classes/query/collection_numbers_test.rb
@@ -1,0 +1,31 @@
+# frozen_string_literal: true
+
+require("test_helper")
+
+# tests of Query::CollectionNumbers class to be included in QueryTest
+module Query::CollectionNumbersTest
+  def test_collection_number_all
+    expects = CollectionNumber.index_order
+    assert_query(expects, :CollectionNumber)
+  end
+
+  def test_collection_number_for_observation
+    obs = observations(:detailed_unknown_obs)
+    expects = CollectionNumber.index_order.for_observation(obs)
+    assert_query(expects, :CollectionNumber, observation: obs.id)
+  end
+
+  def test_collection_number_pattern_search
+    expects = CollectionNumber.index_order.
+              where(CollectionNumber[:name].matches("%Singer%").
+                    or(CollectionNumber[:number].matches("%Singer%"))).
+              sort_by(&:format_name)
+    assert_query(expects, :CollectionNumber, pattern: "Singer")
+
+    expects = CollectionNumber.index_order.
+              where(CollectionNumber[:name].matches("%123a%").
+                    or(CollectionNumber[:number].matches("%123a%"))).
+              sort_by(&:format_name)
+    assert_query(expects, :CollectionNumber, pattern: "123a")
+  end
+end
