+++ test/classes/query/rss_logs_test.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+
+require("test_helper")
+
+# tests of Query::RssLogs class to be included in QueryTest
+module Query::RssLogsTest
+  def test_rss_log_all
+    ids = RssLog.index_order
+    assert_query(ids, :RssLog)
+  end
+
+  def test_rss_log_type
+    ids = [rss_logs(:species_list_rss_log).id]
+    assert_query(ids, :RssLog, type: :species_list)
+  end
+
+  def test_rss_log_in_set
+    rsslog_set_ids = [rss_logs(:species_list_rss_log).id,
+                      rss_logs(:name_rss_log).id]
+    assert_query(rsslog_set_ids, :RssLog, ids: rsslog_set_ids)
+  end
+end
