+++ test/classes/query_test.rb
@@ -1140,2306 +1140,26 @@ class QueryTest < UnitTestCase
   #
   ##############################################################################
 
-  def test_article_all
-    expects = Article.index_order
-    assert_query(expects, :Article)
-  end
-
-  def test_article_by_rss_log
-    assert_query(Article.order_by_rss_log, :Article, by: :rss_log)
-  end
-
-  def test_article_in_set
-    assert_query([articles(:premier_article).id], :Article,
-                 ids: [articles(:premier_article).id])
-    assert_query([], :Article, ids: [])
-  end
-
-  def test_collection_number_all
-    expects = CollectionNumber.index_order
-    assert_query(expects, :CollectionNumber)
-  end
-
-  def test_collection_number_for_observation
-    obs = observations(:detailed_unknown_obs)
-    expects = CollectionNumber.index_order.for_observation(obs)
-    assert_query(expects, :CollectionNumber, observation: obs.id)
-  end
-
-  def test_collection_number_pattern_search
-    expects = CollectionNumber.index_order.
-              where(CollectionNumber[:name].matches("%Singer%").
-                    or(CollectionNumber[:number].matches("%Singer%"))).
-              sort_by(&:format_name)
-    assert_query(expects, :CollectionNumber, pattern: "Singer")
-
-    expects = CollectionNumber.index_order.
-              where(CollectionNumber[:name].matches("%123a%").
-                    or(CollectionNumber[:number].matches("%123a%"))).
-              sort_by(&:format_name)
-    assert_query(expects, :CollectionNumber, pattern: "123a")
-  end
-
-  def test_comment_all
-    expects = Comment.index_order
-    assert_query(expects, :Comment)
-  end
-
-  def test_comment_by_user
-    expects = Comment.index_order.where(user_id: mary.id).distinct
-    assert_query(expects, :Comment, by_user: mary)
-  end
-
-  def test_comment_for_target
-    obs = observations(:minimal_unknown_obs)
-    expects = Comment.index_order.where(target_id: obs.id).distinct
-    assert_query(expects, :Comment, target: obs, type: "Observation")
-  end
-
-  def test_comment_for_user
-    expects = Comment.index_order.select { |c| c.target.user == mary }
-    # expects = Comment.index_order.joins(:target).
-    #           where(targets: { user_id: mary.id }).uniq
-    assert_query(expects, :Comment, for_user: mary)
-    assert_query([], :Comment, for_user: rolf)
-  end
-
-  def test_comment_in_set
-    assert_query([comments(:detailed_unknown_obs_comment).id,
-                  comments(:minimal_unknown_obs_comment_1).id],
-                 :Comment,
-                 ids: [comments(:detailed_unknown_obs_comment).id,
-                       comments(:minimal_unknown_obs_comment_1).id])
-  end
-
-  def test_comment_pattern_search
-    expects = Comment.index_order.
-              where(Comment[:summary].matches("%unknown%").
-                    or(Comment[:comment].matches("%unknown%"))).uniq
-    assert_query(expects, :Comment, pattern: "unknown")
-  end
-
-  def test_external_link_all
-    assert_query(ExternalLink.all.sort_by(&:url), :ExternalLink)
-    assert_query(ExternalLink.where(user: users(:mary)).sort_by(&:url),
-                 :ExternalLink, users: users(:mary))
-    assert_query([], :ExternalLink, users: users(:dick))
-    obs = observations(:coprinus_comatus_obs)
-    assert_query(obs.external_links.sort_by(&:url),
-                 :ExternalLink, observations: obs)
-    obs = observations(:detailed_unknown_obs)
-    assert_query([], :ExternalLink, observations: obs)
-    site = external_sites(:mycoportal)
-    assert_query(site.external_links.sort_by(&:url),
-                 :ExternalLink, external_sites: site)
-    site = external_sites(:inaturalist)
-    assert_query(site.external_links.sort_by(&:url),
-                 :ExternalLink, external_sites: site)
-    assert_query(site.external_links.sort_by(&:url),
-                 :ExternalLink, url: "iNaturalist")
-  end
-
-  # In the model these are all getting a default_scope order: :code thrown on.
-  def test_field_slip_all
-    expects = FieldSlip.index_order
-    assert_query(expects, :FieldSlip)
-  end
-
-  def test_field_slip_by_user
-    expects = FieldSlip.index_order.by_user(mary)
-    assert_query(expects, :FieldSlip, by_user: mary)
-  end
-
-  def test_field_slip_for_project
-    expects = FieldSlip.index_order.where(project: projects(:eol_project))
-    assert_query(expects, :FieldSlip, project: projects(:eol_project))
-  end
-
-  def test_glossary_term_all
-    expects = GlossaryTerm.index_order
-    assert_query(expects, :GlossaryTerm)
-  end
-
-  def test_glossary_term_pattern_search
-    assert_query([], :GlossaryTerm, pattern: "no glossary term has this")
-    # name
-    expects = GlossaryTerm.index_order.
-              where(GlossaryTerm[:description].matches("%conic_glossary_term%").
-                    or(GlossaryTerm[:name].matches("%conic_glossary_term%"))).
-              distinct
-    assert_query(expects, :GlossaryTerm, pattern: "conic_glossary_term")
-    # description
-    expects = GlossaryTerm.index_order.
-              where(GlossaryTerm[:description].matches("%Description of Term%").
-                    or(GlossaryTerm[:name].matches("%Description of Term%"))).
-              distinct
-    assert_query(expects, :GlossaryTerm, pattern: "Description of Term")
-    # blank
-    expects = GlossaryTerm.index_order
-    assert_query(expects, :GlossaryTerm, pattern: "")
-  end
-
-  def test_herbarium_all
-    expects = Herbarium.index_order
-    assert_query(expects.select(:id).distinct, :Herbarium)
-  end
-
-  def test_herbarium_by_records
-    expects = Herbarium.left_outer_joins(:herbarium_records).group(:id).
-              # Wrap known safe argument in Arel
-              # to prevent "Dangerous query method" Deprecation Warning
-              order(HerbariumRecord[:id].count.desc, Herbarium[:id].desc)
-    assert_query(expects, :Herbarium, by: :records)
-  end
-
-  def test_herbarium_in_set
-    expects = [
-      herbaria(:dick_herbarium),
-      herbaria(:nybg_herbarium)
-    ]
-    assert_query(expects, :Herbarium, ids: expects)
-  end
-
-  def test_herbarium_pattern_search
-    # [herbaria(:nybg_herbarium)]
-    expects = Herbarium.index_order.where(
-      Herbarium[:code].concat(Herbarium[:name]).
-      concat(Herbarium[:description].coalesce("")).
-      concat(Herbarium[:mailing_address].coalesce("")).matches("%awesome%")
-    ).distinct
-
-    assert_query(expects, :Herbarium, pattern: "awesome")
-  end
-
-  def test_herbarium_record_all
-    expects = HerbariumRecord.index_order
-    assert_query(expects, :HerbariumRecord)
-  end
-
-  def test_herbarium_record_for_observation
-    obs = observations(:coprinus_comatus_obs)
-    expects = HerbariumRecord.index_order.for_observation(obs)
-    assert_query(expects, :HerbariumRecord, observation: obs.id)
-  end
-
-  def test_herbarium_record_in_herbarium
-    nybg = herbaria(:nybg_herbarium)
-    expects = HerbariumRecord.index_order.where(herbarium: nybg)
-    assert_query(expects, :HerbariumRecord, herbarium: nybg.id)
-  end
-
-  def test_herbarium_record_pattern_search_notes
-    expects = herbarium_record_pattern_search("dried")
-    assert_query(expects, :HerbariumRecord, pattern: "dried")
-  end
-
-  def test_herbarium_record_pattern_search_not_findable
-    assert_query([], :HerbariumRecord,
-                 pattern: "no herbarium record has this")
-  end
-
-  def test_herbarium_record_pattern_search_initial_det
-    expects = herbarium_record_pattern_search("Agaricus")
-    assert_query(expects, :HerbariumRecord, pattern: "Agaricus")
-  end
-
-  def test_herbarium_record_pattern_search_accession_number
-    expects = herbarium_record_pattern_search("123a")
-    assert_query(expects, :HerbariumRecord, pattern: "123a")
-  end
-
-  def test_herbarium_record_pattern_search_blank
-    expects = HerbariumRecord.index_order
-    assert_query(expects, :HerbariumRecord, pattern: "")
-  end
-
-  def herbarium_record_pattern_search(pattern)
-    HerbariumRecord.index_order.where(
-      HerbariumRecord[:initial_det].concat(HerbariumRecord[:accession_number]).
-      concat(HerbariumRecord[:notes].coalesce("")).matches("%#{pattern}%")
-    ).distinct
-  end
-
-  def test_image_all
-    expects = Image.index_order
-    assert_query(expects, :Image)
-  end
-
-  def test_image_for_observations
-    obs = observations(:two_img_obs)
-    expects = Image.index_order.joins(:observations).
-              where(observations: { id: obs.id }).distinct
-    assert_query(expects, :Image, observations: obs)
-  end
-
-  def test_image_for_projects
-    project = projects(:bolete_project)
-    expects = Image.index_order.joins(:projects).
-              where(projects: { id: project.id }).distinct
-    assert_query(expects, :Image, projects: [project.title])
-  end
-
-  def test_image_by_user
-    expects = Image.index_order.where(user_id: rolf.id).distinct
-    assert_query(expects, :Image, by_user: rolf)
-    expects = Image.index_order.where(user_id: mary.id).distinct
-    assert_query(expects, :Image, by_user: mary)
-    expects = Image.index_order.where(user_id: dick.id).distinct
-    assert_query(expects, :Image, by_user: dick)
-  end
-
-  def test_image_in_set
-    ids = [images(:turned_over_image).id,
-           images(:agaricus_campestris_image).id,
-           images(:disconnected_coprinus_comatus_image).id]
-    assert_query(ids, :Image, ids: ids)
-  end
-
-  def test_image_inside_observation
-    obs = observations(:detailed_unknown_obs)
-    assert_equal(2, obs.images.length)
-    expects = obs.images.sort_by(&:id)
-    assert_query(expects, :Image,
-                 observation: obs, outer: 1) # (outer is only used by prev/next)
-    obs = observations(:minimal_unknown_obs)
-    assert_equal(0, obs.images.length)
-    assert_query(obs.images, :Image,
-                 observation: obs, outer: 1) # (outer is only used by prev/next)
-  end
-
-  def test_image_for_project
-    project = projects(:bolete_project)
-    expects = Image.index_order.joins(:project_images).
-              where(project_images: { project: project }).reorder(id: :asc)
-    assert_query(expects, :Image, project: project, by: :id)
-    assert_query([], :Image, project: projects(:empty_project))
-  end
-
-  def test_image_advanced_search_name
-    # expects = [] # [images(:agaricus_campestris_image).id]
-    expects = Image.index_order.joins(observations: :name).
-              where(Name[:search_name].matches("%Agaricus%")).distinct
-    assert_query(expects, :Image, name: "Agaricus")
-  end
-
-  def test_image_advanced_search_user_where
-    expects = Image.index_order.joins(:observations).
-              where(Observation[:where].matches("%burbank%")).
-              where(observations: { is_collection_location: true }).distinct
-    assert_query(expects, :Image, user_where: "burbank")
-
-    assert_query([images(:connected_coprinus_comatus_image).id],
-                 :Image, user_where: "glendale")
-  end
-
-  def test_image_advanced_search_user
-    expects = Image.index_order.joins(observations: :user).
-              where(observations: { user: mary }).
-              order(Image[:created_at].desc, Image[:id].desc).uniq
-    assert_query(expects, :Image, user: "mary")
-  end
-
-  def test_image_advanced_search_content
-    assert_query([images(:turned_over_image).id, images(:in_situ_image).id],
-                 :Image, content: "little")
-    assert_query([images(:connected_coprinus_comatus_image).id],
-                 :Image, content: "fruiting")
-  end
-
-  def test_image_advanced_search_combos
-    assert_query([],
-                 :Image, name: "agaricus", user_where: "glendale")
-    assert_query([images(:agaricus_campestris_image).id],
-                 :Image, name: "agaricus", user_where: "burbank")
-    assert_query([images(:turned_over_image).id, images(:in_situ_image).id],
-                 :Image, content: "little", user_where: "burbank")
-  end
-
-  def test_image_pattern_search
-    assert_query([images(:agaricus_campestris_image).id],
-                 :Image, pattern: "agaricus") # name
-    assert_query([images(:agaricus_campestris_image).id,
-                  images(:connected_coprinus_comatus_image).id,
-                  images(:turned_over_image).id,
-                  images(:in_situ_image).id],
-                 :Image, pattern: "bob dob") # copyright holder
-    assert_query(
-      [images(:in_situ_image).id],
-      :Image, pattern: "looked gorilla OR original" # notes
-    )
-    assert_query([images(:agaricus_campestris_image).id,
-                  images(:connected_coprinus_comatus_image).id],
-                 :Image, pattern: "notes some") # notes
-    assert_query(
-      [images(:turned_over_image).id, images(:in_situ_image).id],
-      :Image, pattern: "dobbs -notes" # (c), not notes
-    )
-    assert_query([images(:in_situ_image).id], :Image,
-                 pattern: "DSCN8835") # original filename
-  end
-
-  def test_image_with_observations
-    expects = Image.index_order.includes(:observations).
-              where.not(observations: { thumb_image: nil }).distinct
-    assert_query(expects, :Image, with_observations: true)
-  end
-
-  # Prove that :with_observations param of Image Query works with each
-  # parameter P for which (a) there's no other test of P for
-  # Image, OR (b) P behaves differently in :with_observations than in
-  # all other params of Image Query's.
-
-  ##### date/time parameters #####
-
-  def test_image_with_observations_created_at
-    created_at = observations(:detailed_unknown_obs).created_at
-    expects = Image.index_order.joins(:observations).
-              where(Observation[:created_at] >= created_at).distinct
-    assert_not_empty(expects, "'expect` is broken; it should not be empty")
-    assert_query(expects, :Image,
-                 with_observations: 1, created_at: created_at)
-  end
-
-  def test_image_with_observations_updated_at
-    updated_at = observations(:detailed_unknown_obs).updated_at
-    expects = Image.index_order.joins(:observations).
-              where(Observation[:updated_at] >= updated_at).distinct
-    assert_not_empty(expects, "'expect` is broken; it should not be empty")
-    assert_query(expects, :Image,
-                 with_observations: 1, updated_at: updated_at)
-  end
-
-  def test_image_with_observations_date
-    date = observations(:detailed_unknown_obs).when
-    expects = Image.index_order.joins(:observations).
-              where(Observation[:when] >= date).distinct
-    assert_not_empty(expects, "'expect` is broken; it should not be empty")
-    assert_query(expects, :Image, with_observations: 1, date: date)
-  end
-
-  ##### list/string parameters #####
-
-  def test_image_with_observations_comments_has
-    expects = Image.index_order.joins(observations: :comments).
-              where(Comment[:summary].matches("%give%")).
-              or(Image.index_order.joins(observations: :comments).
-                 where(Comment[:comment].matches("%give%"))).distinct
-    assert_not_empty(expects, "'expect` is broken; it should not be empty")
-    assert_query(expects, :Image,
-                 with_observations: 1, comments_has: "give")
-  end
-
-  def test_image_with_observations_with_notes_fields
-    obs = observations(:substrate_notes_obs) # obs has notes substrate: field
-    # give it some images
-    obs.images = [images(:conic_image), images(:convex_image)]
-    obs.save
-    expects = Image.index_order.joins(:observations).
-              where(Observation[:notes].matches("%:substrate:%")).uniq
-    assert_not_empty(expects, "'expects` is broken; it should not be empty")
-    assert_query(expects, :Image,
-                 with_observations: 1, with_notes_fields: "substrate")
-  end
-
-  def test_image_with_observations_herbaria
-    name = "The New York Botanical Garden"
-    expects = Image.index_order.
-              joins(observations: { herbarium_records: :herbarium }).
-              where(herbaria: { name: name }).distinct
-    assert_not_empty(expects, "'expect` is broken; it should not be empty")
-    assert_query(expects, :Image, with_observations: 1, herbaria: name)
-  end
-
-  def test_image_with_observations_projects
-    project = projects(:bolete_project)
-    expects = Image.index_order.joins(observations: :projects).
-              where(projects: { title: project.title }).distinct
-    assert_not_empty(expects, "'expect` is broken; it should not be empty")
-    assert_query(expects,
-                 :Image, with_observations: 1, projects: [project.title])
-  end
-
-  def test_image_with_observations_users
-    expects = Image.index_order.joins(:observations).
-              where(observations: { user: dick }).distinct
-    assert_not_empty(expects, "'expect` is broken; it should not be empty")
-    assert_query(expects, :Image, with_observations: 1, users: dick)
-  end
-
-  ##### numeric parameters #####
-
-  def test_image_with_observations_bounding_box
-    obs = give_geolocated_observation_some_images
-
-    lat = obs.lat
-    lng = obs.lng
-    expects = Image.index_order.joins(:observations).
-              where(observations: { lat: lat }).
-              where(observations: { lng: lng }).distinct
-    assert_query(
-      expects,
-      :Image,
-      with_observations: 1,
-      north: lat.to_f, south: lat.to_f, west: lat.to_f, east: lat.to_f
-    )
-  end
-
-  def give_geolocated_observation_some_images
-    obs = observations(:unknown_with_lat_lng) # obs has lat/lon
-    # give it some images
-    obs.images = [images(:conic_image), images(:convex_image)]
-    obs.save
-    obs
-  end
-
-  ##### boolean parameters #####
-
-  def test_image_with_observations_with_comments
-    expects = Image.index_order.joins(observations: :comments).distinct
-    assert_not_empty(expects, "'expect` is broken; it should not be empty")
-    assert_query(expects,
-                 :Image, with_observations: 1, with_comments: true)
-  end
-
-  def test_image_with_observations_with_public_lat_lng
-    give_geolocated_observation_some_images
-
-    expects = Image.index_order.joins(:observations).
-              where.not(observations: { lat: false }).distinct
-    assert_not_empty(expects, "'expect` is broken; it should not be empty")
-    assert_query(expects,
-                 :Image, with_observations: 1, with_public_lat_lng: true)
-  end
-
-  def test_image_with_observations_with_name
-    expects = Image.index_order.joins(:observations).
-              where(observations: { name_id: Name.unknown }).distinct
-    assert_not_empty(expects, "'expect` is broken; it should not be empty")
-    assert_query(expects, :Image, with_observations: 1, with_name: false)
-  end
-
-  def test_image_with_observations_with_notes
-    expects = Image.index_order.joins(:observations).
-              where.not(observations: { notes: Observation.no_notes }).distinct
-    assert_not_empty(expects, "'expect` is broken; it should not be empty")
-    assert_query(expects, :Image, with_observations: 1, with_notes: true)
-  end
-
-  def test_image_with_observations_with_sequences
-    expects = Image.index_order.joins(observations: :sequences).distinct
-    assert_not_empty(expects, "'expect` is broken; it should not be empty")
-    assert_query(expects, :Image, with_observations: 1, with_sequences: true)
-  end
-
-  def test_image_with_observations_is_collection_location
-    expects = Image.index_order.joins(:observations).
-              where(observations: { is_collection_location: true }).distinct
-    assert_not_empty(expects, "'expect` is broken; it should not be empty")
-    assert_query(expects,
-                 :Image, with_observations: 1, is_collection_location: true)
-  end
-
-  def test_image_with_observations_at_location
-    expects = Image.index_order.joins(observations: :location).
-              where(observations: { location: locations(:burbank) }).
-              where(observations: { is_collection_location: true }).distinct
-    assert_query(expects,
-                 :Image, with_observations: 1, location: locations(:burbank).id)
-    assert_query([], :Image,
-                 with_observations: 1, location: locations(:mitrula_marsh).id)
-  end
-
-  def test_image_with_observations_at_where
-    assert_query([images(:connected_coprinus_comatus_image).id],
-                 :Image, with_observations: 1, user_where: "glendale")
-    assert_query([],
-                 :Image, with_observations: 1, user_where: "snazzle")
-  end
-
-  def test_image_with_observations_by_user
-    expects = image_with_observations_by_user(rolf)
-    assert_query(expects.to_a, :Image, with_observations: 1, by_user: rolf)
-
-    expects = image_with_observations_by_user(mary)
-    assert_query(expects.to_a, :Image, with_observations: 1, by_user: mary)
-
-    assert_query([], :Image,
-                 with_observations: 1, by_user: users(:zero_user))
-  end
-
-  def image_with_observations_by_user(user)
-    Image.index_order.joins(:observations).
-      where(observations: { user: user }).distinct
-  end
-
-  def test_image_with_observations_for_project
-    assert_query([],
-                 :Image,
-                 with_observations: 1, project: projects(:empty_project))
-    assert_query(observations(:two_img_obs).images.index_order.distinct,
-                 :Image,
-                 with_observations: 1, project: projects(:two_img_obs_project))
-  end
-
-  def test_image_with_observations_in_set
-    obs_ids = [observations(:detailed_unknown_obs).id,
-               observations(:agaricus_campestris_obs).id]
-    # There's an order_by find_in_set thing here we can't do in Arel.
-    # But luckily we can just quote the method.
-    oids = obs_ids.join(",")
-    expects = Image.joins(:observations).where(observations: { id: obs_ids }).
-              reorder(Arel.sql("FIND_IN_SET(observations.id,'#{oids}')").asc,
-                      Image[:id].desc).distinct
-    assert_query(expects, :Image, with_observations: 1, obs_ids: obs_ids)
-    assert_query([], :Image,
-                 with_observations: 1,
-                 obs_ids: [observations(:minimal_unknown_obs).id])
-  end
-
-  def test_image_with_observations_in_species_list
-    assert_query([images(:turned_over_image).id,
-                  images(:in_situ_image).id],
-                 :Image,
-                 with_observations: 1,
-                 species_list: species_lists(:unknown_species_list).id)
-    assert_query([], :Image,
-                 with_observations: 1,
-                 species_list: species_lists(:first_species_list).id)
-  end
-
-  def test_image_with_observations_of_children
-    assert_query([images(:agaricus_campestris_image).id],
-                 :Image,
-                 with_observations: 1,
-                 names: [names(:agaricus).id], include_subtaxa: true)
-  end
-
-  def test_image_sorted_by_original_name
-    assert_query([images(:turned_over_image).id,
-                  images(:connected_coprinus_comatus_image).id,
-                  images(:disconnected_coprinus_comatus_image).id,
-                  images(:in_situ_image).id,
-                  images(:commercial_inquiry_image).id,
-                  images(:agaricus_campestris_image).id],
-                 :Image,
-                 ids: [images(:in_situ_image).id,
-                       images(:turned_over_image).id,
-                       images(:commercial_inquiry_image).id,
-                       images(:disconnected_coprinus_comatus_image).id,
-                       images(:connected_coprinus_comatus_image).id,
-                       images(:agaricus_campestris_image).id],
-                 by: :original_name)
-  end
-
-  def test_image_with_observations_of_name
-    expects = Image.index_order.joins(:observation_images, :observations).
-              where(observations: { name: names(:fungi) }).distinct
-    assert_query(expects,
-                 :Image, with_observations: 1, names: [names(:fungi).id])
-    assert_query([images(:connected_coprinus_comatus_image).id],
-                 :Image,
-                 with_observations: 1, names: [names(:coprinus_comatus).id])
-    assert_query([images(:agaricus_campestris_image).id],
-                 :Image,
-                 with_observations: 1, names: [names(:agaricus_campestris).id])
-    assert_query([], :Image,
-                 with_observations: 1, names: [names(:conocybe_filaris).id])
-  end
-
-  def test_location_all
-    expects = Location.index_order
-    assert_query(expects, :Location)
-    expects = Location.reorder(id: :asc)
-    assert_query(expects, :Location, by: :id)
-  end
-
-  def test_location_by_user
-    assert_query(Location.reorder(id: :asc).where(user: rolf).distinct,
-                 :Location, by_user: rolf, by: :id)
-    assert_query([], :Location, by_user: users(:zero_user))
-  end
-
-  def test_location_by_editor
-    assert_query([], :Location, by_editor: rolf)
-    User.current = mary
-    loc = Location.where.not(user: mary).index_order.first
-    loc.display_name = "new name"
-    loc.save
-    assert_query([loc], :Location, by_editor: mary)
-    assert_query([], :Location, by_editor: dick)
-  end
-
-  def test_location_by_rss_log
-    expects = Location.order_by_rss_log
-    assert_query(expects.to_a, :Location, by: :rss_log)
-  end
-
-  def test_location_in_set
-    assert_query([locations(:gualala).id,
-                  locations(:albion).id,
-                  locations(:burbank).id,
-                  locations(:elgin_co).id],
-                 :Location,
-                 ids: [locations(:gualala).id,
-                       locations(:albion).id,
-                       locations(:burbank).id,
-                       locations(:elgin_co).id])
-  end
-
-  def test_location_pattern_search
-    expects = Location.reorder(id: :asc).
-              where(Location[:name].matches("%California%")).distinct
-    assert_query(expects, :Location, pattern: "California", by: :id)
-    assert_query([locations(:elgin_co).id],
-                 :Location, pattern: "Canada")
-    assert_query([], :Location, pattern: "Canada -Elgin")
-  end
-
-  def test_location_advanced_search_name
-    assert_query([locations(:burbank).id],
-                 :Location, name: "agaricus")
-    assert_query([], :Location, name: "coprinus")
-  end
-
-  def test_location_advanced_search_user_where
-    assert_query([locations(:burbank).id],
-                 :Location, user_where: "burbank")
-    assert_query([locations(:howarth_park).id,
-                  locations(:salt_point).id],
-                 :Location, user_where: "park")
-  end
-
-  def test_location_advanced_search_user
-    expects = Location.index_order.joins(observations: :user).
-              where(observations: { user: rolf }).distinct
-    assert_query(expects, :Location, user: "rolf")
-
-    expects = Location.index_order.joins(observations: :user).
-              where(observations: { user: dick }).distinct
-    assert_query(expects, :Location, user: "dick")
-  end
-
-  def test_location_advanced_search_content
-    # content in obs.notes
-    assert_query([locations(:burbank).id],
-                 :Location, content: '"strange place"')
-    # content in Comment
-    assert_query(
-      [locations(:burbank).id],
-      :Location, content: '"a little of everything"'
-    )
-    # no search loc.notes
-    assert_query([],
-                 :Location, content: '"play with"')
-  end
-
-  def test_location_advanced_search_content_combos
-    assert_query([locations(:burbank).id],
-                 :Location, name: "agaricus", content: '"lawn"')
-    assert_query([],
-                 :Location, name: "agaricus", content: '"play with"')
-    # from observation and comment for same observation
-    assert_query([locations(:burbank).id],
-                 :Location,
-                 content: '"a little of everything" "strange place"')
-    # from different comments, should fail
-    assert_query([],
-                 :Location,
-                 content: '"minimal unknown" "complicated"')
-  end
-
-  def test_location_regexp_search
-    expects = Location.where(Location[:name].matches_regexp("California")).
-              index_order.distinct
-    assert_query(expects, :Location, regexp: ".alifornia")
-  end
-
-  def test_location_with_descriptions
-    expects = Location.joins(:descriptions).index_order.distinct
-    assert_query(expects, :Location, with_descriptions: 1)
-  end
-
-  def test_location_with_descriptions_by_user
-    expects = Location.joins(:descriptions).
-              where(descriptions: { user: rolf }).index_order.distinct
-    assert_query(expects, :Location, with_descriptions: 1, by_user: rolf)
-
-    assert_query([], :Location, with_descriptions: 1, by_user: mary)
-  end
-
-  def test_location_with_descriptions_by_author
-    expects = Location.joins(descriptions: :location_description_authors).
-              where(location_description_authors: { user: rolf }).
-              index_order.distinct
-    assert_query(expects, :Location, with_descriptions: 1, by_author: rolf)
-    assert_query([], :Location, with_descriptions: 1, by_author: mary)
-  end
-
-  def test_location_with_descriptions_by_editor
-    User.current = mary
-    desc = location_descriptions(:albion_desc)
-    desc.notes = "blah blah blah"
-    desc.save
-    assert_query([], :Location, with_descriptions: 1, by_editor: rolf)
-
-    expects = Location.joins(descriptions: :location_description_editors).
-              where(location_description_editors: { user: mary }).
-              index_order.distinct
-    assert_query(expects, :Location, with_descriptions: 1, by_editor: mary)
-  end
-
-  def test_location_with_descriptions_in_set
-    assert_query(
-      [locations(:albion), locations(:no_mushrooms_location)],
-      :Location,
-      with_descriptions: 1,
-      desc_ids: [location_descriptions(:albion_desc).id,
-                 location_descriptions(:no_mushrooms_location_desc).id]
-    )
-    assert_query([locations(:albion)], :Location,
-                 with_descriptions: 1,
-                 desc_ids: [location_descriptions(:albion_desc).id, rolf.id])
-    assert_query([],
-                 :Location, with_descriptions: 1, desc_ids: [rolf.id])
-  end
-
-  def test_location_with_observations
-    expects = Location.joins(:observations).index_order.distinct
-    assert_query(expects, :Location, with_observations: 1)
-  end
-
-  # Prove that :with_observations param of Location Query works with each
-  # parameter P for which (a) there's no other test of P for
-  # Location, OR (b) P behaves differently in :with_observations than in
-  # all other params of Location Query's.
-
-  ##### date/time parameters #####
-
-  def test_location_with_observations_created_at
-    created_at = observations(:california_obs).created_at
-    expects = Location.index_order.joins(:observations).
-              where(Observation[:created_at] >= created_at).distinct
-    assert_query(
-      expects, :Location, with_observations: 1, created_at: created_at
-    )
-  end
-
-  def test_location_with_observations_updated_at
-    updated_at = observations(:california_obs).updated_at
-    expects = Location.index_order.joins(:observations).
-              where(Observation[:updated_at] >= updated_at).distinct
-    assert_query(
-      expects, :Location, with_observations: 1, updated_at: updated_at
-    )
-  end
-
-  def test_location_with_observations_date
-    date = observations(:california_obs).when
-    expects = Location.index_order.joins(:observations).
-              where(Observation[:when] >= date).distinct
-    assert_query(
-      expects, :Location, with_observations: 1, date: date
-    )
-  end
-
-  ##### list/string parameters #####
-
-  def test_location_with_observations_include_subtaxa
-    parent = names(:agaricus)
-    children = Name.index_order.
-               where(Name[:text_name].matches_regexp(parent.text_name))
-    assert_query(
-      Location.joins(:observations).
-               where(observations: { name: [parent] + children }).distinct,
-      :Location,
-      with_observations: 1, names: parent.text_name, include_subtaxa: true
-    )
-  end
-
-  def test_location_with_observations_comments_has
-    # Create a Comment, unfortunately hitting the db because
-    # (a) the query should find multiple Locations;
-    # (b) all Observation fixtures with Comments have the same Location; and
-    # (c) adding a Comment fixture breaks the tests.
-    Comment.create(
-      user: katrina,
-      summary: "Another cool obs",
-      comment: "with different location than minimal_unknown_obs",
-      target_type: Observation,
-      target: observations(:vouchered_obs)
-    )
-    assert_query(
-      Location.index_order.joins(observations: :comments).
-               where(Comment[:summary].matches("%cool%")).
-               or(
-                 Location.index_order.joins(observations: :comments).
-                          where(Comment[:comment].matches("%cool%"))
-               ).distinct,
-      :Location, with_observations: 1, comments_has: "cool"
-    )
-  end
-
-  def test_location_with_observations_with_notes_fields
-    assert_query(
-      Location.index_order.joins(:observations).
-               where(Observation[:notes].matches("%:substrate:%")).distinct,
-      :Location, with_observations: 1, with_notes_fields: "substrate"
-    )
-  end
-
-  def test_location_with_observations_herbaria
-    name = "The New York Botanical Garden"
-    expects = Location.joins(observations: { herbarium_records: :herbarium }).
-              where(herbaria: { name: name }).index_order.distinct
-    assert_query(expects, :Location, with_observations: 1, herbaria: name)
-  end
-
-  def test_location_with_observations_names
-    names = [names(:boletus_edulis), names(:agaricus_campestris)].
-            map(&:text_name)
-    expects = Location.joins(observations: :name).
-              where(observations: { text_name: names }).index_order.distinct
-    assert_query(expects, :Location, with_observations: 1, names: names)
-  end
-
-  def test_location_with_observations_notes_has
-    expects = Location.index_order.joins(:observations).
-              where(Observation[:notes].matches("%somewhere%")).distinct
-    assert_query(
-      expects, :Location, with_observations: 1, notes_has: "somewhere"
-    )
-  end
-
-  def test_location_with_observations_locations
-    loc_with_observations = locations(:burbank)
-    loc_without_observations = locations(:no_mushrooms_location)
-    locations = [loc_with_observations, loc_without_observations]
-    assert_query(
-      [loc_with_observations],
-      :Location, with_observations: 1, locations: locations.map(&:name)
-    )
-  end
-
-  def test_location_with_observations_projects
-    project = projects(:bolete_project)
-    assert_query(
-      Location.index_order.joins(observations: :projects).
-               where(projects: { title: project.title }).distinct,
-      :Location, with_observations: 1, projects: project.title
-    )
-  end
-
-  def test_location_with_observations_include_synonyms
-    # Create Observations of synonyms, unfortunately hitting the db, because:
-    # (a) there are no Observation fixtures for a Name with a synonym_names; and
-    # (b) tests are brittle, so adding an Observation fixture will break them.
-    Observation.create(
-      location: locations(:albion),
-      name: names(:macrolepiota_rachodes),
-      user: rolf
-    )
-    Observation.create(
-      location: locations(:howarth_park),
-      name: names(:macrolepiota_rhacodes),
-      user: rolf
-    )
-    assert_query(
-      [locations(:albion), locations(:howarth_park)],
-      :Location,
-      with_observations: 1, names: "Macrolepiota rachodes",
-      include_synonyms: true
-    )
-  end
-
-  def test_location_with_observations_users
-    assert_query(
-      Location.index_order.joins(:observations).
-      where(observations: { user: dick }).distinct,
-      :Location, with_observations: 1, users: dick
-    )
-  end
-
-  ##### numeric parameters #####
-
-  def test_location_with_observations_confidence
-    # Create Observations with both vote_cache and location, because:
-    # (a) there aren't Observation fixtures like that, and
-    # (b) tests are brittle, so adding or modifying a fixture will break them.
-    obses = Observation.index_order.where(vote_cache: 1..3)
-    obses.each { |obs| obs.update!(location: locations(:albion)) }
-    expects = Location.index_order.joins(:observations).
-              where(observations: { vote_cache: 1..3 }).distinct
-    assert_not_empty(expects, "'expect` is broken; it should not be empty")
-    assert_query(expects, :Location,
-                 with_observations: 1, confidence: [1.0, 3.0])
-  end
-
-  ##### boolean parameters #####
-  def test_location_with_observations_with_comments
-    assert_query(
-      Location.index_order.joins(observations: :comments).distinct,
-      :Location, with_observations: 1, with_comments: true
-    )
-  end
-
-  def test_location_with_observations_with_public_lat_lng
-    assert_query(
-      Location.joins(:observations).where(observations: { gps_hidden: false }).
-               where.not(observations: { lat: false }).index_order.distinct,
-      :Location, with_observations: 1, with_public_lat_lng: true
-    )
-  end
-
-  def test_location_with_observations_with_name
-    expects = Location.index_order.joins(:observations).
-              where(observations: { name: Name.unknown }).distinct
-    assert_query(expects, :Location, with_observations: 1, with_name: false)
-  end
-
-  def test_location_with_observations_with_notes
-    expects = Location.index_order.joins(:observations).
-              where.not(observations: { notes: Observation.no_notes }).distinct
-    assert_query(expects, :Location, with_observations: 1, with_notes: true)
-  end
-
-  def test_location_with_observations_with_sequences
-    expects = Location.index_order.joins(observations: :sequences).distinct
-    assert_query(expects, :Location, with_observations: 1, with_sequences: true)
-  end
-
-  def test_location_with_observations_is_collection_location
-    expects = Location.index_order.joins(:observations).
-              where(observations: { is_collection_location: true }).distinct
-    assert_query(
-      expects, :Location, with_observations: 1, is_collection_location: true
-    )
-  end
-
-  def test_location_with_observations_by_user
-    expects = location_with_observations_by_user(rolf)
-    assert_query(expects, :Location, with_observations: 1, by_user: rolf.id)
-
-    zero_user = users(:zero_user)
-    expects = location_with_observations_by_user(zero_user)
-    assert_equal(0, expects.length)
-    assert_query(expects, :Location, with_observations: 1, by_user: zero_user)
-  end
-
-  def location_with_observations_by_user(user)
-    Location.joins(:observations).where(observations: { user: user }).
-      index_order.distinct
-  end
-
-  def test_location_with_observations_for_project
-    assert_query([],
-                 :Location,
-                 with_observations: 1, project: projects(:empty_project))
-    assert_query([observations(:collected_at_obs).location],
-                 :Location,
-                 with_observations: 1,
-                 project: projects(:obs_collected_and_displayed_project))
-  end
-
-  def test_location_with_observations_in_set
-    assert_query([locations(:burbank).id],
-                 :Location,
-                 with_observations: 1,
-                 obs_ids: [observations(:minimal_unknown_obs).id])
-    assert_query([], :Location,
-                 with_observations: 1,
-                 obs_ids: [observations(:coprinus_comatus_obs).id])
-  end
-
-  def test_location_with_observations_in_species_list
-    assert_query([locations(:burbank).id],
-                 :Location,
-                 with_observations: 1,
-                 species_list: species_lists(:unknown_species_list).id)
-    assert_query([], :Location,
-                 with_observations: 1,
-                 species_list: species_lists(:first_species_list).id)
-  end
-
-  def test_location_with_observations_of_children
-    assert_query([locations(:burbank).id],
-                 :Location,
-                 with_observations: 1,
-                 names: [names(:agaricus).id], include_subtaxa: true)
-  end
-
-  def test_location_with_observations_of_name
-    assert_query([locations(:burbank).id], :Location,
-                 with_observations: 1,
-                 names: [names(:agaricus_campestris).id])
-    assert_query([], :Location,
-                 with_observations: 1,
-                 names: [names(:peltigera).id])
-  end
-
-  def test_location_description_all
-    gualala = locations(:gualala)
-    all_descs = LocationDescription.all.to_a
-    all_gualala_descs = LocationDescription.
-                        where(location: gualala).to_a
-    public_gualala_descs = LocationDescription.
-                           where(location: gualala, public: true).to_a
-    assert(all_gualala_descs.length < all_descs.length)
-    assert(public_gualala_descs.length < all_gualala_descs.length)
-
-    assert_query(all_descs, :LocationDescription, by: :id)
-    assert_query(all_gualala_descs, :LocationDescription,
-                 by: :id, locations: gualala)
-    assert_query(public_gualala_descs, :LocationDescription,
-                 by: :id, locations: gualala, public: "yes")
-  end
-
-  def test_location_description_by_user
-    expects = LocationDescription.where(user: rolf).to_a
-    assert_query(expects, :LocationDescription, by_user: rolf)
-
-    expects = LocationDescription.where(user: mary).to_a
-    assert_equal(0, expects.length)
-    assert_query(expects, :LocationDescription, by_user: mary)
-  end
-
-  def test_location_description_by_author
-    loc1, loc2, loc3 = Location.all.index_order
-    desc1 =
-      loc1.description ||= LocationDescription.create!(location_id: loc1.id)
-    desc2 =
-      loc2.description ||= LocationDescription.create!(location_id: loc2.id)
-    desc3 =
-      loc3.description ||= LocationDescription.create!(location_id: loc3.id)
-    desc1.add_author(rolf)
-    desc2.add_author(mary)
-    desc3.add_author(rolf)
-
-    # Using Rails instead of db; don't know how to do it with .joins & .where
-    descs = LocationDescription.all
-    assert_query(descs.find_all { |d| d.authors.include?(rolf) },
-                 :LocationDescription, by_author: rolf, by: :id)
-    assert_query(descs.find_all { |d| d.authors.include?(mary) },
-                 :LocationDescription, by_author: mary)
-    assert_query([], :LocationDescription, by_author: users(:zero_user))
-  end
-
-  def test_location_description_by_editor
-    loc1, loc2, loc3 = Location.index_order
-    desc1 =
-      loc1.description ||= LocationDescription.create!(location_id: loc1.id)
-    desc2 =
-      loc2.description ||= LocationDescription.create!(location_id: loc2.id)
-    desc3 =
-      loc3.description ||= LocationDescription.create!(location_id: loc3.id)
-    desc1.add_editor(rolf) # Fails since he's already an author!
-    desc2.add_editor(mary)
-    desc3.add_editor(rolf)
-
-    # Using Rails instead of db; don't know how to do it with .joins & .where
-    descs = LocationDescription.all
-    assert_query(descs.find_all { |d| d.editors.include?(rolf) },
-                 :LocationDescription, by_editor: rolf, by: :id)
-    assert_query(descs.find_all { |d| d.editors.include?(mary) },
-                 :LocationDescription, by_editor: mary)
-    assert_query([], :LocationDescription, by_editor: users(:zero_user))
-  end
-
-  def test_location_description_in_set
-    assert_query([],
-                 :LocationDescription,
-                 ids: rolf.id)
-    assert_query(LocationDescription.all,
-                 :LocationDescription,
-                 ids: LocationDescription.select(:id).to_a)
-    assert_query([location_descriptions(:albion_desc).id],
-                 :LocationDescription,
-                 ids: [rolf.id, location_descriptions(:albion_desc).id])
-  end
-
-  def test_location_description_coercion
-    ds1 = location_descriptions(:albion_desc)
-    ds2 = location_descriptions(:no_mushrooms_location_desc)
-    description_coercion_assertions(ds1, ds2, :Location)
-  end
-
-  def test_name_all
-    # NOTE: misspellings are modified by `do_test_name_all`
-    # This saves looking up Name.index_order a bunch of times.
-    expect = Name.index_order
-    expects = expect.to_a
-    # SQL does not sort 'Kuhner' and 'Kühner'
-    do_test_name_all(expect) if sql_collates_accents?
-
-    pair = expects.select { |x| x.text_name == "Lentinellus ursinus" }
-    a = expects.index(pair.first)
-    b = expects.index(pair.last)
-    expects[a], expects[b] = expects[b], expects[a]
-    do_test_name_all(expect)
-  end
-
-  def do_test_name_all(expect)
-    expect_good = expect.with_correct_spelling
-    expect_bad  = expect.with_incorrect_spelling
-    assert_query(expect_good.to_a, :Name)
-    assert_query(expect.to_a, :Name, misspellings: :either)
-    assert_query(expect_good.to_a, :Name, misspellings: :no)
-    assert_query(expect_bad.to_a, :Name, misspellings: :only)
-  end
-
-  def test_name_by_user
-    assert_query(Name.index_order.where(user: mary).with_correct_spelling,
-                 :Name, by_user: mary)
-    assert_query(Name.index_order.where(user: dick).with_correct_spelling,
-                 :Name, by_user: dick)
-    assert_query(Name.index_order.where(user: rolf).with_correct_spelling,
-                 :Name, by_user: rolf)
-    assert_query([], :Name, by_user: users(:zero_user))
-  end
-
-  def test_name_by_editor
-    assert_query([], :Name, by_editor: rolf)
-    assert_query([], :Name, by_editor: mary)
-    expects = Name.reorder(id: :asc).
-              with_correct_spelling.by_editor(dick).distinct
-    assert_query(expects, :Name, by_editor: dick, by: :id)
-  end
-
-  def test_name_by_rss_log
-    expects = Name.order_by_rss_log
-    assert_query(expects, :Name, by: :rss_log)
-  end
-
-  def test_name_in_set
-    assert_query([names(:fungi).id,
-                  names(:coprinus_comatus).id,
-                  names(:conocybe_filaris).id,
-                  names(:lepiota_rhacodes).id,
-                  names(:lactarius_subalpinus).id],
-                 :Name,
-                 ids: [names(:fungi).id,
-                       names(:coprinus_comatus).id,
-                       names(:conocybe_filaris).id,
-                       names(:lepiota_rhacodes).id,
-                       names(:lactarius_subalpinus).id])
-  end
-
-  def test_name_include_subtaxa_exclude_original_names
-    expects = Name.index_order.subtaxa_of(names(:agaricus))
-    assert_query(expects, :Name,
-                 names: [names(:agaricus).id], include_subtaxa: true,
-                 exclude_original_names: true)
-  end
-
-  def test_name_need_description
-    expects = Name.index_order.description_needed.distinct
-    assert_query(expects, :Name, need_description: 1)
-  end
-
-  def test_name_rank_single
-    expects = Name.with_correct_spelling.with_rank("Family").index_order
-    assert_query(expects, :Name, rank: "Family")
-  end
-
-  # NOTE: Something is wrong in the fixtures between Genus and Family
-  def test_name_rank_range
-    expects = Name.with_correct_spelling.
-              with_rank_between("Genus", "Kingdom").index_order
-    assert_query(expects, :Name, rank: %w[Genus Kingdom])
-
-    expects = Name.with_correct_spelling.with_rank("Family").index_order
-    assert_query(expects, :Name, rank: %w[Family Family])
-  end
-
-  def test_name_pattern_search
-    assert_query(
-      [],
-      :Name, pattern: "petigera" # search_name
-    )
-    assert_query(
-      [names(:petigera).id],
-      :Name, pattern: "petigera", misspellings: :either
-    )
-    assert_query(
-      [names(:peltigera).id],
-      :Name, pattern: "ye auld manual of lichenes" # citation
-    )
-    assert_query(
-      [names(:agaricus_campestras).id],
-      :Name, pattern: "prevent me" # description notes
-    )
-    assert_query(
-      [names(:suillus)],
-      :Name, pattern: "smell as sweet" # gen_desc
-    )
-    # Prove pattern search gets hits for description look_alikes
-    assert_query(
-      [names(:peltigera).id],
-      :Name, pattern: "superficially similar"
-    )
-  end
-
-  def test_name_advanced_search
-    assert_query([names(:macrocybe_titans).id], :Name,
-                 name: "macrocybe*titans")
-    assert_query([names(:coprinus_comatus).id], :Name,
-                 user_where: "glendale") # where
-    expects = Name.index_order.joins(:observations).
-              where(Observation[:location_id].eq(locations(:burbank).id)).
-              distinct
-    assert_query(expects, :Name, user_where: "burbank") # location
-    expects = Name.index_order.joins(:observations).
-              where(Observation[:user_id].eq(rolf.id)).distinct
-    assert_query(expects, :Name, user: "rolf")
-    assert_query([names(:coprinus_comatus).id], :Name,
-                 content: "second fruiting") # notes
-    assert_query([names(:fungi).id], :Name,
-                 content: '"a little of everything"') # comment
-  end
-
-  def test_name_with_descriptions
-    expects = Name.index_order.with_correct_spelling.
-              joins(:descriptions).distinct
-    assert_query(expects, :Name, with_descriptions: 1)
-  end
-
-  def test_name_with_descriptions_by_user
-    expects = name_with_descriptions_by_user(mary)
-    assert_query(expects, :Name, with_descriptions: 1, by_user: mary)
-
-    expects = name_with_descriptions_by_user(dick)
-    assert_query(expects, :Name, with_descriptions: 1, by_user: dick)
-  end
-
-  def name_with_descriptions_by_user(user)
-    Name.with_correct_spelling.joins(:descriptions).
-      where(name_descriptions: { user: user }).index_order.distinct
-  end
-
-  def test_name_with_descriptions_by_author
-    expects = name_with_descriptions_by_author(rolf)
-    assert_query(expects, :Name, with_descriptions: 1, by_author: rolf)
-
-    expects = name_with_descriptions_by_author(mary)
-    assert_query(expects, :Name, with_descriptions: 1, by_author: mary)
-
-    expects = name_with_descriptions_by_author(dick)
-    assert_query(expects, :Name, with_descriptions: 1, by_author: dick)
-  end
-
-  def name_with_descriptions_by_author(user)
-    Name.with_correct_spelling.
-      joins(descriptions: :name_description_authors).
-      where(name_description_authors: { user: user }).index_order.distinct
-  end
-
-  def test_name_with_descriptions_by_editor
-    expects = name_with_descriptions_by_editor(rolf)
-    assert_query(expects, :Name, with_descriptions: 1, by_editor: rolf)
-
-    expects = name_with_descriptions_by_editor(rolf)
-    assert_query(expects, :Name, with_descriptions: 1, by_editor: mary)
-
-    expects = name_with_descriptions_by_editor(dick)
-    assert_equal(0, expects.length)
-    assert_query(expects, :Name, with_descriptions: 1, by_editor: dick)
-  end
-
-  def name_with_descriptions_by_editor(user)
-    Name.with_correct_spelling.
-      joins(descriptions: :name_description_editors).
-      where(name_description_editors: { user: user }).index_order.distinct
-  end
-
-  def test_name_with_descriptions_in_set
-    desc1 = name_descriptions(:peltigera_desc)
-    desc2 = name_descriptions(:peltigera_alt_desc)
-    desc3 = name_descriptions(:draft_boletus_edulis)
-    name1 = names(:peltigera)
-    name2 = names(:boletus_edulis)
-    assert_query([name1, name2],
-                 :Name,
-                 with_descriptions: 1, desc_ids: [desc1, desc2, desc3])
-  end
-
-  def test_name_with_observations
-    expects = Name.with_correct_spelling.joins(:observations).
-              select(:name).distinct.pluck(:name_id).sort
-    assert_query(expects, :Name, with_observations: 1, by: :id)
-  end
-
-  # Prove that :with_observations param of Name Query works with each
-  # parameter P for which (a) there's no other test of P for
-  # Name, OR (b) P behaves differently in :with_observations than in
-  # all other params of Name Query's.
-
-  ##### date/time parameters #####
-
-  def test_name_with_observations_created_at
-    created_at = observations(:california_obs).created_at
-    expects = Name.index_order.with_correct_spelling.joins(:observations).
-              where(Observation[:created_at] >= created_at).distinct
-    assert_query(expects, :Name, with_observations: 1, created_at: created_at)
-  end
-
-  def test_name_with_observations_updated_at
-    updated_at = observations(:california_obs).updated_at
-    expects = Name.index_order.with_correct_spelling.joins(:observations).
-              where(Observation[:updated_at] >= updated_at).distinct
-    assert_query(expects, :Name, with_observations: 1, updated_at: updated_at)
-  end
-
-  def test_name_with_observations_date
-    date = observations(:california_obs).when
-    expects = Name.index_order.with_correct_spelling.joins(:observations).
-              where(Observation[:when] >= date).distinct
-    assert_query(expects, :Name, with_observations: 1, date: date)
-  end
-
-  ##### list/string parameters #####
-
-  def test_name_with_observations_with_notes_fields
-    expects = Name.index_order.with_correct_spelling.joins(:observations).
-              where(Observation[:notes].matches("%:substrate:%")).distinct
-    assert_query(
-      expects, :Name, with_observations: 1, with_notes_fields: "substrate"
-    )
-  end
-
-  def test_name_with_observations_herbaria
-    name = "The New York Botanical Garden"
-    expects = Name.index_order.with_correct_spelling.
-              joins(observations: { herbarium_records: :herbarium }).
-              where(herbaria: { name: name }).distinct
-    assert_query(expects, :Name, with_observations: 1, herbaria: name)
-  end
-
-  def test_name_with_observations_projects
-    project = projects(:bolete_project)
-    expects = Name.index_order.with_correct_spelling.
-              joins({ observations: :project_observations }).
-              where(project_observations: { project: project }).distinct
-    # project.observations.map(&:name).uniq
-    assert_query(expects, :Name, with_observations: 1, projects: project.title)
-  end
-
-  def test_name_with_observations_users
-    expects = Name.index_order.with_correct_spelling.joins(:observations).
-              where(observations: { user: dick }).distinct
-    assert_query(expects, :Name, with_observations: 1, users: dick)
-  end
-
-  ##### numeric parameters #####
-
-  def test_name_with_observations_confidence
-    expects = Name.index_order.with_correct_spelling.joins(:observations).
-              where(observations: { vote_cache: 1..3 }).distinct
-    assert_not_empty(expects, "'expect` is broken; it should not be empty")
-    assert_query(expects, :Name, with_observations: 1, confidence: [1, 3])
-
-    # north/south/east/west
-    obs = observations(:unknown_with_lat_lng)
-    lat = obs.lat
-    lng = obs.lng
-    expects = Name.index_order.with_correct_spelling.joins(:observations).
-              where(observations: { lat: lat, lng: lng }).distinct
-    assert_query(
-      expects,
-      :Name,
-      with_observations: 1,
-      north: lat.to_f, south: lat.to_f, west: lat.to_f, east: lat.to_f
-    )
-  end
-
-  ##### boolean parameters #####
-
-  def test_name_with_observations_with_comments
-    expects = Name.index_order.with_correct_spelling.
-              joins(observations: :comments).distinct
-    assert_query(expects, :Name, with_observations: 1, with_comments: true)
-  end
-
-  def test_name_with_observations_with_public_lat_lng
-    expects = Name.index_order.joins(:observations).
-              where.not(observations: { lat: false }).distinct
-    assert_query(
-      expects, :Name, with_observations: 1, with_public_lat_lng: true
-    )
-  end
-
-  def test_name_with_observations_with_name
-    expects = Name.index_order.with_correct_spelling.joins(:observations).
-              where(observations: { name_id: Name.unknown }).distinct
-    assert_query(expects, :Name, with_observations: 1, with_name: false)
-  end
-
-  def test_name_with_observations_with_notes
-    expects = Name.index_order.with_correct_spelling.joins(:observations).
-              where.not(observations: { notes: Observation.no_notes }).distinct
-    assert_query(expects, :Name, with_observations: 1, with_notes: true)
-  end
-
-  def test_name_with_observations_with_sequences
-    expects = Name.index_order.with_correct_spelling.
-              joins(observations: :sequences).distinct
-    assert_query(expects, :Name, with_observations: 1, with_sequences: true)
-  end
-
-  def test_name_with_observations_is_collection_location
-    expects = Name.index_order.with_correct_spelling.joins(:observations).
-              where(observations: { is_collection_location: true }).distinct
-    assert_query(
-      expects, :Name, with_observations: 1, is_collection_location: true
-    )
-  end
-
-  def test_name_with_observations_at_location
-    loc = locations(:burbank)
-    expects = Name.index_order.with_correct_spelling.joins(:observations).
-              where(observations: { location: loc }).distinct
-    assert_query(expects, :Name, with_observations: 1, location: loc)
-  end
-
-  def test_name_with_observations_at_where
-    assert_query([names(:coprinus_comatus).id],
-                 :Name, with_observations: 1, user_where: "glendale")
-  end
-
-  def test_name_with_observations_by_user
-    assert_query(name_with_observations_by_user(rolf),
-                 :Name, with_observations: 1, by_user: rolf)
-    assert_query(name_with_observations_by_user(mary),
-                 :Name, with_observations: 1, by_user: mary)
-    assert_query([], :Name, with_observations: 1, by_user: users(:zero_user))
-  end
-
-  def name_with_observations_by_user(user)
-    Name.index_order.with_correct_spelling.joins(:observations).
-      where(observations: { user: user }).distinct
-  end
-
-  def test_name_with_observations_for_project
-    project = projects(:empty_project)
-    assert_query([], :Name, with_observations: 1, project: project)
-
-    project2 = projects(:two_img_obs_project)
-    expects = Name.index_order.with_correct_spelling.
-              joins({ observations: :project_observations }).
-              where(project_observations: { project: project2 }).distinct
-    assert_query(expects, :Name, with_observations: 1, project: project2)
-  end
-
-  def test_name_with_observations_in_set
-    oids = three_amigos.join(",")
-    expects = Name.with_correct_spelling.joins(:observations).
-              where(observations: { id: three_amigos }).
-              reorder(
-                Arel.sql("FIND_IN_SET(observations.id,'#{oids}')").asc,
-                id: :desc
-              ).distinct
-    assert_query(expects, :Name, with_observations: 1, obs_ids: three_amigos)
-  end
-
-  def test_name_with_observations_in_species_list
-    spl = species_lists(:unknown_species_list)
-    expects = Name.index_order.with_correct_spelling.
-              joins({ observations: :species_list_observations }).
-              where(species_list_observations: { species_list: spl }).uniq
-    assert_query(expects, :Name, with_observations: 1, species_list: spl)
-
-    spl2 = species_lists(:first_species_list)
-    assert_query([], :Name, with_observations: 1, species_list: spl2)
-  end
-
-  def test_name_description_all
-    pelt = names(:peltigera)
-    all_descs = NameDescription.all.to_a
-    all_pelt_descs = NameDescription.where(name: pelt).to_a
-    public_pelt_descs = NameDescription.where(name: pelt, public: true).to_a
-    assert(all_pelt_descs.length < all_descs.length)
-    assert(public_pelt_descs.length < all_pelt_descs.length)
-
-    assert_query(all_descs, :NameDescription, by: :id)
-    assert_query(all_pelt_descs, :NameDescription, by: :id, names: pelt)
-    assert_query(public_pelt_descs, :NameDescription,
-                 by: :id, names: pelt, public: "yes")
-  end
-
-  def test_name_description_by_user
-    expects = NameDescription.where(user: mary).order(:id)
-    assert_query(expects, :NameDescription, by_user: mary, by: :id)
-
-    expects = NameDescription.where(user: katrina).order(:id)
-    assert_query(expects, :NameDescription, by_user: katrina, by: :id)
-
-    assert_query([], :NameDescription, by_user: junk, by: :id)
-  end
-
-  def test_name_description_by_author
-    expects = NameDescription.joins(:name_description_authors).
-              where(name_description_authors: { user_id: rolf }).order(:id)
-    assert_query(expects, :NameDescription, by_author: rolf, by: :id)
-
-    expects = NameDescription.joins(:name_description_authors).
-              where(name_description_authors: { user_id: mary }).order(:id)
-    assert_query(expects, :NameDescription, by_author: mary, by: :id)
-
-    assert_query([], :NameDescription, by_author: junk)
-  end
-
-  def test_name_description_by_editor
-    expects = NameDescription.joins(:name_description_editors).
-              where(name_description_editors: { user_id: rolf }).order(:id)
-    assert_query(expects, :NameDescription, by_editor: rolf)
-
-    expects = NameDescription.joins(:name_description_editors).
-              where(name_description_editors: { user_id: mary }).order(:id)
-    assert_query(expects, :NameDescription, by_editor: mary)
-
-    assert_query([], :NameDescription, by_editor: dick)
-  end
-
-  def test_name_description_in_set
-    assert_query([],
-                 :NameDescription, ids: rolf.id)
-    assert_query(NameDescription.all,
-                 :NameDescription, ids: NameDescription.select(:id).to_a)
-    assert_query([NameDescription.first.id],
-                 :NameDescription,
-                 ids: [rolf.id, NameDescription.first.id])
-  end
-
-  def test_name_description_with_default_desc
-    assert_query(NameDescription.is_default.index_order,
-                 :NameDescription, with_default_desc: 1)
-    assert_query(NameDescription.is_not_default.index_order,
-                 :NameDescription, with_default_desc: 0)
-  end
-
-  def test_name_description_desc_type_user
-    assert_query(NameDescription.where(source_type: 5).index_order,
-                 :NameDescription, desc_type: "user")
-  end
-
-  def test_name_description_desc_type_project
-    assert_query(NameDescription.where(source_type: 3).index_order,
-                 :NameDescription, desc_type: "project")
-  end
-
-  def test_name_description_desc_project
-    assert_query(NameDescription.
-                 where(project: projects(:eol_project)).index_order,
-                 :NameDescription, desc_project: projects(:eol_project).id)
-  end
-
-  def test_name_description_desc_creator
-    assert_query(NameDescription.where(user: rolf).index_order,
-                 :NameDescription, desc_creator: rolf.id)
-    assert_query(NameDescription.where(user: mary).index_order,
-                 :NameDescription, desc_creator: mary.id)
-  end
-
-  # waiting on a new AbstractModel scope for searches,
-  # plus a specific NameDescription scope coalescing the fields.
-  # def test_name_description_desc_content
-  #   assert_query(NameDescription.notes_has("some notes")).index_order,
-  #                :NameDescription, desc_content: "some notes")
-  # end
-
-  def test_name_description_ok_for_export
-    assert_query(NameDescription.where(ok_for_export: 1).index_order,
-                 :NameDescription, ok_for_export: 1)
-    assert_query(NameDescription.where(ok_for_export: 0).index_order,
-                 :NameDescription, ok_for_export: 0)
-  end
-
-  def test_observation_all
-    expects = Observation.index_order
-    assert_query(expects, :Observation)
-  end
-
-  def test_observation_in_project_list
-    project = projects(:bolete_project)
-    # expects = project.species_lists.map(&:observations).flatten.to_a
-    expects = Observation.index_order.
-              joins(species_lists: :project_species_lists).
-              where(project_species_lists: { project: project }).distinct
-    assert_query(expects, :Observation, project_lists: project.title)
-  end
-
-  def test_observation_at_location
-    expects = Observation.index_order.
-              where(location: locations(:burbank)).distinct
-    assert_query(expects, :Observation, location: locations(:burbank))
-  end
-
-  # Overwrites scope `order_by_rss_log` in abstract_model
-  def test_observation_by_rss_log
-    expects = Observation.order_by_rss_log
-    assert_query(expects, :Observation, by: :rss_log)
-  end
-
-  def test_observation_by_user
-    expects = Observation.reorder(id: :asc).where(user: rolf.id).to_a
-    assert_query(expects, :Observation, by_user: rolf, by: :id)
-    expects = Observation.reorder(id: :asc).where(user: mary.id).to_a
-    assert_query(expects, :Observation, by_user: mary, by: :id)
-    expects = Observation.reorder(id: :asc).where(user: dick.id).to_a
-    assert_query(expects, :Observation, by_user: dick, by: :id)
-    assert_query([], :Observation, by_user: junk, by: :id)
-  end
-
-  def test_observation_for_project
-    assert_query([],
-                 :Observation, project: projects(:empty_project))
-    assert_query(projects(:bolete_project).observations,
-                 :Observation, project: projects(:bolete_project))
-  end
-
-  def test_observation_for_project_projects_equivalence
-    qu1 = Query.lookup_and_save(:Observation,
-                                project: projects(:bolete_project))
-    qu2 = Query.lookup_and_save(:Observation,
-                                projects: projects(:bolete_project).id.to_s)
-    assert_equal(qu1.results, qu2.results)
-  end
-
-  def test_observation_in_set
-    obs_set_ids = [observations(:unknown_with_no_naming).id,
-                   observations(:minimal_unknown_obs).id,
-                   observations(:strobilurus_diminutivus_obs).id,
-                   observations(:detailed_unknown_obs).id,
-                   observations(:agaricus_campestros_obs).id,
-                   observations(:coprinus_comatus_obs).id,
-                   observations(:agaricus_campestras_obs).id,
-                   observations(:agaricus_campestris_obs).id,
-                   observations(:agaricus_campestrus_obs).id]
-    assert_query(obs_set_ids, :Observation, ids: obs_set_ids)
-  end
-
-  def test_observation_in_species_list
-    # These two are identical, so should be disambiguated by reverse_id.
-    assert_query([observations(:detailed_unknown_obs).id,
-                  observations(:minimal_unknown_obs).id],
-                 :Observation,
-                 species_list: species_lists(:unknown_species_list).id)
-  end
-
-  def test_observation_of_children
-    name = names(:agaricus)
-    expects = Observation.index_order.
-              of_name(name, include_subtaxa: true).distinct
-    assert_query(expects, :Observation, names: [name.id], include_subtaxa: true)
-  end
-
-  def test_observation_of_name_with_modifiers
-    User.current = rolf
-    expects = Observation.index_order.where(name: names(:fungi)).distinct
-    assert_query(expects, :Observation, names: [names(:fungi).id])
-    assert_query([],
-                 :Observation, names: [names(:macrolepiota_rachodes).id])
-
-    # test all truthy/falsy combinations of these boolean parameters:
-    #  include_synonyms, include_all_name_proposals, exclude_consensus
-    names = Name.index_order.
-            where(Name[:text_name].matches("Agaricus camp%")).to_a
-    agaricus_ssp = names.clone
-    name = names.pop
-    names.each { |n| name.merge_synonyms(n) }
-    observations(:agaricus_campestras_obs).update(user: mary)
-    observations(:agaricus_campestros_obs).update(user: mary)
-
-    # observations where name(s) is consensus
-    assert_query([observations(:agaricus_campestris_obs).id],
-                 :Observation,
-                 names: [names(:agaricus_campestris).id],
-                 include_synonyms: false,
-                 include_all_name_proposals: false,
-                 exclude_consensus: false)
-
-    # name(s) is consensus, but is not the consensus (an oxymoron)
-    assert_query([],
-                 :Observation,
-                 names: [names(:agaricus_campestris).id],
-                 include_synonyms: false,
-                 include_all_name_proposals: false,
-                 exclude_consensus: true)
-
-    # name(s) is proposed
-    assert_query([observations(:agaricus_campestris_obs).id,
-                  observations(:coprinus_comatus_obs).id],
-                 :Observation,
-                 names: [names(:agaricus_campestris).id],
-                 include_synonyms: false,
-                 include_all_name_proposals: true,
-                 exclude_consensus: false)
-
-    # name(s) is proposed, but is not the consensus
-    assert_query([observations(:coprinus_comatus_obs).id],
-                 :Observation,
-                 names: [names(:agaricus_campestris).id],
-                 include_synonyms: false,
-                 include_all_name_proposals: true,
-                 exclude_consensus: true)
-
-    # consensus is a synonym of name(s)
-    assert_query([observations(:agaricus_campestros_obs).id,
-                  observations(:agaricus_campestras_obs).id,
-                  observations(:agaricus_campestrus_obs).id,
-                  observations(:agaricus_campestris_obs).id],
-                 :Observation,
-                 names: [names(:agaricus_campestris).id],
-                 include_synonyms: true,
-                 include_all_name_proposals: false,
-                 exclude_consensus: false)
-
-    # same as above but exclude_original_names
-    # conensus is a synonym of name(s) other than name(s)
-    assert_query([observations(:agaricus_campestros_obs).id,
-                  observations(:agaricus_campestras_obs).id,
-                  observations(:agaricus_campestrus_obs).id],
-                 :Observation,
-                 names: [names(:agaricus_campestris).id],
-                 include_synonyms: true,
-                 exclude_original_names: true)
-
-    # consensus is a synonym of name(s) but not a synonym of name(s) (oxymoron)
-    assert_query([],
-                 :Observation,
-                 names: [names(:agaricus_campestras).id],
-                 include_synonyms: true,
-                 include_all_name_proposals: false,
-                 exclude_consensus: true)
-
-    # where synonyms of names are proposed
-    assert_query([observations(:agaricus_campestros_obs).id,
-                  observations(:agaricus_campestras_obs).id,
-                  observations(:agaricus_campestrus_obs).id,
-                  observations(:agaricus_campestris_obs).id,
-                  observations(:coprinus_comatus_obs).id],
-                 :Observation,
-                 names: [names(:agaricus_campestris).id],
-                 include_synonyms: true,
-                 include_all_name_proposals: true,
-                 exclude_consensus: false)
-
-    # where synonyms of name are proposed, but are not the consensus
-    assert_query([observations(:coprinus_comatus_obs).id],
-                 :Observation,
-                 names: [names(:agaricus_campestras).id],
-                 include_synonyms: true,
-                 include_all_name_proposals: true,
-                 exclude_consensus: true)
-
-    spl = species_lists(:first_species_list)
-    spl.observations << observations(:agaricus_campestrus_obs)
-    spl.observations << observations(:agaricus_campestros_obs)
-    proj = projects(:eol_project)
-    proj.observations << observations(:agaricus_campestris_obs)
-    proj.observations << observations(:agaricus_campestras_obs)
-
-    assert_query([observations(:agaricus_campestros_obs).id,
-                  observations(:agaricus_campestrus_obs).id],
-                 :Observation,
-                 names: agaricus_ssp.map(&:text_name),
-                 species_lists: [spl.title])
-
-    assert_query([observations(:agaricus_campestras_obs).id,
-                  observations(:agaricus_campestris_obs).id],
-                 :Observation,
-                 names: agaricus_ssp.map(&:text_name),
-                 projects: [proj.title])
-  end
-
-  def test_observation_pattern_search
-    # notes
-    # assert_query([observations(:agaricus_campestras_obs).id,
-    #               observations(:agaricus_campestros_obs).id,
-    #               observations(:agaricus_campestrus_obs).id,
-    #               observations(:strobilurus_diminutivus_obs).id],
-    #              :Observation, pattern: '"somewhere else"')
-    # where
-    assert_query([observations(:strobilurus_diminutivus_obs).id],
-                 :Observation, pattern: "pipi valley")
-    # location
-    expects = observation_pattern_search("burbank")
-    assert_query(expects.uniq, :Observation, pattern: "burbank", by: :name)
-
-    # name
-    expects = observation_pattern_search("agaricus")
-    assert_query(expects.uniq, :Observation, pattern: "agaricus", by: :name)
-  end
-
-  def observation_pattern_search(pattern)
-    Observation.joins(:name).
-      where(Name[:search_name].concat(Observation[:where]).
-            matches("%#{pattern}%")).
-      reorder(Name[:sort_name].asc,
-              Observation[:when].desc, Observation[:id].desc)
-  end
-
-  def test_observation_advanced_search
-    assert_query([observations(:strobilurus_diminutivus_obs).id],
-                 :Observation, name: "diminutivus")
-    assert_query([observations(:coprinus_comatus_obs).id],
-                 :Observation, user_where: "glendale") # where
-    expects = Observation.reorder(id: :asc).
-              where(location: locations(:burbank)).distinct
-    assert_query(expects, :Observation,
-                 user_where: "burbank", by: :id) # location
-    expects = Observation.reorder(id: :asc).where(user: rolf.id).distinct
-    assert_query(expects, :Observation, user: "rolf", by: :id)
-    assert_query([observations(:coprinus_comatus_obs).id], # notes
-                 :Observation, content: "second fruiting")
-    assert_query([observations(:minimal_unknown_obs).id],
-                 :Observation, content: "agaricus") # comment
-  end
-
-  def test_observation_date
-    # blank should return all
-    assert_query(Observation.index_order, :Observation, date: nil)
-    # impossible dates should return none
-    assert_query([], :Observation, date: %w[1550 1551])
-    # single date should return after
-    assert_query(Observation.index_order.when_after("2011-05-12"),
-                 :Observation, date: "2011-05-12")
-    # year should return after
-    assert_query(Observation.index_order.when_after("2005"),
-                 :Observation, date: "2005")
-    # years should return between
-    assert_query(Observation.index_order.when_between("2005", "2009"),
-                 :Observation, date: %w[2005 2009])
-    # in a month range, any year
-    assert_query(Observation.index_order.when_between("05", "12"),
-                 :Observation, date: %w[05 12])
-    # in a date range, any year
-    assert_query(Observation.index_order.when_between("02-22", "08-22"),
-                 :Observation, date: %w[02-22 08-22])
-    # period wraps around the new year
-    assert_query(Observation.index_order.when_between("08-22", "02-22"),
-                 :Observation, date: %w[08-22 02-22])
-    # full dates
-    assert_query(Observation.index_order.
-                 when_between("2009-08-22", "2009-10-20"),
-                 :Observation, date: %w[2009-08-22 2009-10-20])
-    # date wraps around the new year
-    assert_query(Observation.index_order.
-                 when_between("2015-08-22", "2016-02-22"),
-                 :Observation, date: %w[2015-08-22 2016-02-22])
-  end
-
-  def test_observation_created_at
-    # blank should return all
-    assert_query(Observation.index_order, :Observation, created_at: nil)
-    # impossible dates should return none
-    assert_query([], :Observation, created_at: %w[2000 2001])
-    # single datetime should return after
-    assert_query(Observation.index_order.
-                 created_after("2011-05-12-12-59-57"),
-                 :Observation, created_at: "2011-05-12-12-59-57")
-    # single date should return after
-    assert_query(Observation.index_order.created_after("2011-05-12"),
-                 :Observation, created_at: "2011-05-12")
-    # year should return after
-    assert_query(Observation.index_order.created_after("2005"),
-                 :Observation, created_at: "2005")
-    # years should return between
-    assert_query(Observation.index_order.created_between("2005", "2009"),
-                 :Observation, created_at: %w[2005 2009])
-    # full dates
-    assert_query(Observation.index_order.
-                 created_between("2009-08-22", "2009-10-20"),
-                 :Observation, created_at: %w[2009-08-22 2009-10-20])
-    # full datetimes
-    assert_query(Observation.index_order.
-                 created_between("2009-08-22-03-04-22", "2009-10-20-03-04-22"),
-                 :Observation,
-                 created_at: %w[2009-08-22-03-04-22 2009-10-20-03-04-22])
-  end
-
-  def test_observation_updated_at
-    # blank should return all
-    assert_query(Observation.index_order, :Observation, updated_at: nil)
-    # impossible dates should return none
-    assert_query([], :Observation, updated_at: %w[2000 2001])
-    # single datetime should return after
-    assert_query(Observation.index_order.
-                 updated_after("2011-05-12-12-59-57"),
-                 :Observation, updated_at: "2011-05-12-12-59-57")
-    # single date should return after
-    assert_query(Observation.index_order.updated_after("2011-05-12"),
-                 :Observation, updated_at: "2011-05-12")
-    # year should return after
-    assert_query(Observation.index_order.updated_after("2005"),
-                 :Observation, updated_at: "2005")
-    # years should return between
-    assert_query(Observation.index_order.updated_between("2005", "2009"),
-                 :Observation, updated_at: %w[2005 2009])
-    # full dates
-    assert_query(Observation.index_order.
-                 updated_between("2009-08-22", "2009-10-20"),
-                 :Observation, updated_at: %w[2009-08-22 2009-10-20])
-    # full datetimes
-    assert_query(Observation.index_order.
-                 updated_between("2009-08-22-03-04-22", "2009-10-20-03-04-22"),
-                 :Observation,
-                 updated_at: %w[2009-08-22-03-04-22 2009-10-20-03-04-22])
-  end
-
-  def test_project_all
-    expects = Project.index_order
-    assert_query(expects, :Project)
-  end
-
-  def test_project_by_rss_log
-    expects = Project.order_by_rss_log
-    assert_query(expects, :Project, by: :rss_log)
-  end
-
-  def test_project_in_set
-    assert_query([projects(:eol_project).id], :Project,
-                 ids: [projects(:eol_project).id])
-    assert_query([], :Project, ids: [])
-  end
-
-  def test_project_pattern_search
-    assert_query([],
-                 :Project, pattern: "no project has this")
-    # title
-    expects = project_pattern_search("bolete")
-    assert_query(expects, :Project, pattern: "bolete")
-    # summary
-    expects = project_pattern_search("two lists")
-    assert_query(expects, :Project, pattern: "two lists")
-
-    expects = Project.index_order
-    assert_query(expects, :Project, pattern: "")
-  end
-
-  def project_pattern_search(pattern)
-    Project.where(Project[:title].matches("%#{pattern}%").
-                  or(Project[:summary].matches("%#{pattern}%"))).
-      index_order.distinct
-  end
-
-  def test_rss_log_all
-    ids = RssLog.index_order
-    assert_query(ids, :RssLog)
-  end
-
-  def test_rss_log_type
-    ids = [rss_logs(:species_list_rss_log).id]
-    assert_query(ids, :RssLog, type: :species_list)
-  end
-
-  def test_rss_log_in_set
-    rsslog_set_ids = [rss_logs(:species_list_rss_log).id,
-                      rss_logs(:name_rss_log).id]
-    assert_query(rsslog_set_ids, :RssLog, ids: rsslog_set_ids)
-  end
-
-  def test_sequence_all
-    expects = Sequence.index_order
-    assert_query(expects, :Sequence)
-  end
-
-  def test_sequence_locus_has
-    assert_query(Sequence.where(Sequence[:locus].matches("ITS%")).
-                 index_order.distinct,
-                 :Sequence, locus_has: "ITS")
-  end
-
-  def test_sequence_archive
-    assert_query([sequences(:alternate_archive)],
-                 :Sequence, archive: "UNITE")
-  end
-
-  def test_sequence_accession_has
-    assert_query([sequences(:deposited_sequence)],
-                 :Sequence, accession_has: "968605")
-  end
-
-  def test_sequence_notes_has
-    assert_query([sequences(:deposited_sequence)],
-                 :Sequence, notes_has: "deposited_sequence")
-  end
-
-  def test_sequence_for_observations
-    obs = observations(:locally_sequenced_obs)
-    assert_query([sequences(:local_sequence)],
-                 :Sequence, observations: [obs.id])
-  end
-
-  def test_sequence_filters
-    sequences = Sequence.reorder(id: :asc).all
-    seq1 = sequences[0]
-    seq2 = sequences[1]
-    seq3 = sequences[3]
-    seq4 = sequences[4]
-    seq1.update(observation: observations(:minimal_unknown_obs))
-    seq2.update(observation: observations(:detailed_unknown_obs))
-    seq3.update(observation: observations(:agaricus_campestris_obs))
-    seq4.update(observation: observations(:peltigera_obs))
-    assert_query([seq1, seq2], :Sequence, obs_date: %w[2006 2006])
-    assert_query([seq1, seq2], :Sequence, observers: users(:mary))
-    assert_query([seq1, seq2], :Sequence, names: "Fungi")
-    assert_query([seq4], :Sequence,
-                 names: "Petigera", include_synonyms: true)
-    expects = Sequence.index_order.joins(:observation).
-              where(observations: { location: locations(:burbank) }).
-              or(Sequence.index_order.joins(:observation).
-                 where(Observation[:where].matches("Burbank"))).distinct
-    assert_query(expects, :Sequence, locations: "Burbank")
-    assert_query([seq2], :Sequence, projects: "Bolete Project")
-    assert_query([seq1, seq2], :Sequence,
-                 species_lists: "List of mysteries")
-    assert_query([seq4], :Sequence, confidence: "2")
-    # The test returns these sequences in random order, can't work.
-    # assert_query([seq1, seq2, seq3], :Sequence,
-    #              north: "90", south: "0", west: "-180", east: "-100")
-  end
-
-  def test_uses_join_hash
-    query = Query.lookup(:Sequence,
-                         north: "90", south: "0", west: "-180", east: "-100")
-    assert_not(query.uses_join_sub([], :location))
-    assert(query.uses_join_sub([:location], :location))
-    assert_not(query.uses_join_sub({}, :location))
-    assert(query.uses_join_sub({ test: :location }, :location))
-    assert(query.uses_join_sub(:location, :location))
-  end
-
-  def test_sequence_in_set
-    list_set_ids = [sequences(:fasta_formatted_sequence).id,
-                    sequences(:bare_formatted_sequence).id]
-    assert_query(list_set_ids, :Sequence, ids: list_set_ids)
-  end
-
-  def test_sequence_pattern_search
-    assert_query([], :Sequence, pattern: "nonexistent")
-    assert_query(Sequence.where(Sequence[:locus].matches("ITS%")).
-                 index_order.distinct,
-                 :Sequence, pattern: "ITS")
-    assert_query([sequences(:alternate_archive)],
-                 :Sequence, pattern: "UNITE")
-    assert_query([sequences(:deposited_sequence)],
-                 :Sequence, pattern: "deposited_sequence")
-  end
-
-  def test_species_list_all
-    expects = SpeciesList.index_order
-    assert_query(expects, :SpeciesList)
-  end
-
-  def test_species_list_sort_by_user
-    expects = SpeciesList.order_by_user.to_a
-    assert_query(expects, :SpeciesList, by: :user)
-  end
-
-  def test_species_list_sort_by_title
-    expects = SpeciesList.order(:title).to_a
-    assert_query(expects, :SpeciesList, by: :title)
-  end
-
-  def test_species_list_at_location
-    expects = SpeciesList.where(location: locations(:burbank)).
-              index_order.distinct
-    assert_query(expects, :SpeciesList, location: locations(:burbank))
-    assert_query(
-      [], :SpeciesList, location: locations(:unused_location)
-    )
-  end
-
-  def test_species_list_at_where
-    assert_query([], :SpeciesList, user_where: "nowhere")
-    assert_query([species_lists(:where_no_mushrooms_list)],
-                 :SpeciesList, user_where: "no mushrooms")
-  end
-
-  def test_species_list_by_rss_log
-    assert_query(SpeciesList.order_by_rss_log, :SpeciesList, by: :rss_log)
-  end
-
-  def test_species_list_by_user
-    expects = SpeciesList.where(user: mary).index_order.distinct
-    assert_query(expects, :SpeciesList, by_user: mary)
-    assert_query([], :SpeciesList, by_user: dick)
-  end
-
-  def test_species_list_by_user_sort_by_id
-    expects = SpeciesList.where(user: rolf).reorder(id: :asc).uniq
-    assert_query(expects, :SpeciesList, by_user: rolf, by: :id)
-  end
-
-  def test_species_list_for_project
-    assert_query([],
-                 :SpeciesList, project: projects(:empty_project))
-    assert_query(projects(:bolete_project).species_lists,
-                 :SpeciesList, project: projects(:bolete_project))
-    assert_query(
-      projects(:two_list_project).species_lists,
-      :SpeciesList, project: projects(:two_list_project)
-    )
-  end
-
-  def test_species_list_in_set
-    list_set_ids = [species_lists(:first_species_list).id,
-                    species_lists(:unknown_species_list).id]
-    assert_query(list_set_ids, :SpeciesList, ids: list_set_ids)
-  end
-
-  def test_species_list_pattern_search
-    assert_query([],
-                 :SpeciesList, pattern: "nonexistent pattern")
-    # in title
-    pattern = "query_first_list"
-    expects = species_list_pattern_search(pattern)
-    assert_query(expects, :SpeciesList, pattern: "query_first_list")
-    # in notes
-    pattern = species_lists(:query_notes_list).notes
-    expects = species_list_pattern_search(pattern)
-    assert_query(expects, :SpeciesList, pattern: pattern)
-    # in location
-    pattern = locations(:burbank).name
-    expects = species_list_pattern_search(pattern)
-    assert_query(expects, :SpeciesList, pattern: locations(:burbank).name)
-    # in where
-    pattern = species_lists(:where_list).where
-    expects = species_list_pattern_search(pattern)
-    assert_query(expects, :SpeciesList, pattern: pattern)
-
-    expects = SpeciesList.index_order
-    assert_query(expects, :SpeciesList, pattern: "")
-  end
-
-  def species_list_pattern_search(pattern)
-    SpeciesList.index_order.left_outer_joins(:location).
-      where(SpeciesList[:title].matches("%#{pattern}%").
-            or(SpeciesList[:notes].matches("%#{pattern}%")).
-            or(SpeciesList[:where].matches("%#{pattern}%")).
-            or(Location[:name].matches("%#{pattern}%"))).distinct
-  end
-
-  def test_user_all
-    expects = User.order(:name).to_a
-    assert_query(expects, :User)
-    expects = User.order(:login).to_a
-    assert_query(expects, :User, by: :login)
-  end
-
-  def test_user_in_set
-    assert_query([rolf.id, mary.id, junk.id], :User,
-                 ids: [junk.id, mary.id, rolf.id], by: :reverse_name)
-  end
-
-  def test_user_pattern_search_nonexistent
-    assert_query([],
-                 :User, pattern: "nonexistent pattern")
-  end
-
-  def test_user_pattern_search_login
-    # in login
-    expects = user_pattern_search(users(:spammer).login)
-    assert_query(expects, :User, pattern: users(:spammer).login)
-  end
-
-  def test_user_pattern_search_name
-    # in name
-    expects = user_pattern_search(users(:mary).name)
-    assert_query(expects, :User, pattern: users(:mary).name)
-  end
-
-  def test_user_pattern_search_blank
-    assert_query(User.order(name: :asc, id: :desc).to_a,
-                 :User, pattern: "")
-  end
-
-  def test_user_pattern_search_sorted_by_location
-    # sorted by location should include Users without location
-    # (Differs from searches on other Classes or by other sort orders)
-    expects = User.left_outer_joins(:location).
-              order(Location[:name].asc, User[:id].desc).uniq
-    assert_query(expects, :User, pattern: "", by: "location")
-  end
-
-  def user_pattern_search(pattern)
-    User.where(User[:login].matches("%#{pattern}%").
-               or(User[:name].matches("%#{pattern}%"))).
-      order(name: :asc, id: :desc).uniq
-  end
-  ##############################################################################
-  #
-  #  :section: Filters
-  #
-  ##############################################################################
-
-  def test_filtering_content_with_images
-    expects = Observation.where.not(thumb_image_id: nil).index_order.uniq
-    assert_query(expects, :Observation, with_images: "yes")
-
-    expects = Observation.where(thumb_image_id: nil).index_order.uniq
-    assert_query(expects, :Observation, with_images: "no")
-  end
-
-  def test_filtering_content_with_specimen
-    expects = Observation.where(specimen: true).index_order.uniq
-    assert_query(expects, :Observation, with_specimen: "yes")
-
-    expects = Observation.where(specimen: false).index_order.uniq
-    assert_query(expects, :Observation, with_specimen: "no")
-  end
-
-  def test_filtering_content_with_lichen
-    expects_obs = Observation.where(Observation[:lifeform].matches("%lichen%")).
-                  index_order.uniq
-    expects_names = Name.with_correct_spelling.
-                    where(Name[:lifeform].matches("%lichen%")).index_order.uniq
-    assert_query(expects_obs, :Observation, lichen: "yes")
-    assert_query(expects_names, :Name, lichen: "yes")
-  end
-
-  def test_filtering_content_with_non_lichen
-    expects_obs = Observation.
-                  where(Observation[:lifeform].does_not_match("% lichen %")).
-                  index_order.uniq
-    expects_names = Name.with_correct_spelling.
-                    where(Name[:lifeform].does_not_match("% lichen %")).
-                    index_order.uniq
-    assert_query(expects_obs, :Observation, lichen: "no")
-    assert_query(expects_names, :Name, lichen: "no")
-  end
-
-  def test_filtering_content_region
-    expects = Location.where(Location[:name].matches("%California%")).
-              index_order.uniq
-    assert_query(expects, :Location, region: "California, USA")
-    assert_query(expects, :Location, region: "USA, California")
-
-    expects = Observation.index_order.
-              where(Observation[:where].matches("%California, USA")).uniq
-    assert_query(expects, :Observation, region: "California, USA")
-
-    expects = Location.where(Location[:name].matches("%, USA").
-              or(Location[:name].matches("%, Canada"))).index_order.uniq
-    assert(expects.include?(locations(:albion))) # usa
-    assert(expects.include?(locations(:elgin_co))) # canada
-    assert_query(expects, :Location, region: "North America")
-  end
-
-  def test_filtering_content_clade
-    names = Name.reorder(id: :asc).
-            with_correct_spelling.where(text_name: "Agaricales").or(
-              Name.reorder(id: :asc).where(
-                Name[:classification].matches_regexp("Order: _Agaricales_")
-              )
-            ).reorder(sort_name: :asc, id: :desc).distinct
-    obs = Observation.reorder(id: :asc).where(text_name: "Agaricales").or(
-      Observation.reorder(id: :asc).where(
-        Observation[:classification].matches_regexp("Order: _Agaricales_")
-      )
-    ).reorder(when: :desc, id: :desc).distinct
-    assert_query(obs, :Observation, clade: "Agaricales")
-    assert_query(names, :Name, clade: "Agaricales")
-  end
+  include Query::ArticlesTest
+  include Query::CollectionNumbersTest
+  include Query::CommentsTest
+  include Query::ExternalLinksTest
+  include Query::FieldSlipsTest
+  include Query::GlossaryTermsTest
+  include Query::HerbariaTest
+  include Query::HerbariumRecordsTest
+  include Query::ImagesTest
+  include Query::LocationDescriptionsTest
+  include Query::LocationsTest
+  include Query::NameDescriptionsTest
+  include Query::NamesTest
+  include Query::ObservationsTest
+  include Query::ProjectsTest
+  include Query::RssLogsTest
+  include Query::SequencesTest
+  include Query::SpeciesListsTest
+  include Query::UsersTest
+  include Query::ContentFiltersTest
 
   ##############################################################################
   #
